components verplaatst en verbeterd
