+++ sca_frontend/sca_frontend/src/components/MatchCard.tsx
@@ -8,9 +8,9 @@ interface MatchCardProps {
 
 const MatchCard: React.FC<MatchCardProps> = ({ details, result, points }) => {
   return (
-    <div className="flex justify-between items-center w-full p-2 bg-black rounded text-white">
-      <span>{details}</span>
-      <span className={`text-${points >= 0 ? 'green-500' : 'red-500'}`}>
+    <div className="flex justify-between items-center p-2 bg-black rounded">
+      <span className="text-white">{details}</span>
+      <span className={points > 0 ? 'text-green-500' : 'text-red-500'}>
         {result} {points > 0 ? `+${points}` : points}
       </span>
     </div>
