+++ sca_frontend/sca_frontend/src/components/MatchHistory.tsx
@@ -7,10 +7,8 @@ interface MatchHistoryProps {
 }
 
 const MatchHistory: React.FC<MatchHistoryProps> = ({ matches, className }) => (
-  <div className={`flex flex-col bg-gray-800 rounded-lg p-4 ${className}`}>
-    
-      <h3 className="text-lg font-semibold mb-4 text-white">Match History</h3> 
-    
+  <div className={`flex flex-col rounded-lg ${className}`}>
+    <h3 className="text-lg font-bold text-black mb-2">Match History</h3>
     <div className="flex flex-col space-y-2">
       {matches.map((match, index) => (
         <MatchCard
